+++ shared/src/chess/board.ts
@@ -1,5 +1,23 @@
 import { Cell } from "./cell";
-import { Color, Piece } from "./piece";
+import { BoardMeta } from "./game";
+import {
+  Bishop,
+  Color,
+  King,
+  Knight,
+  Pawn,
+  Piece,
+  PieceMeta,
+  PieceType,
+  Queen,
+  Rook,
+} from "./piece";
+import {
+  DiagonalMovementRule,
+  HorizontalMovementRule,
+  VerticalMovementRule,
+} from "./rules";
+import { fromChessToLogic } from "./turn-formatter";
 
 export class Board {
   public size = 8;
@@ -10,43 +28,67 @@ export class Board {
     );
   }
 
-  private xCharToIndex = (char: string) => char.charCodeAt(0) - 97;
-  private yCharToIndex = (char: string) => parseInt(char);
-  private isIndexInValid = (index: number) => index >= 0 || index < this.size;
+  public isIndexInValid = (index: number) => index >= 0 || index < this.size;
 
-  move(color: Color, from: string, to: string) {
-    const [fromXChar, fromYChar] = from.split("");
-    const [toXChar, toYChar] = to.split("");
-
-    const fromX = this.xCharToIndex(fromXChar);
-    const toX = this.xCharToIndex(toXChar);
-    const fromY = this.yCharToIndex(fromYChar);
-    const toY = this.yCharToIndex(toYChar);
-
-    if (
-      !(
-        this.isIndexInValid(fromX) &&
-        this.isIndexInValid(toX) &&
-        this.isIndexInValid(toX) &&
-        this.isIndexInValid(toY)
-      )
-    ) {
-      throw new Error("Invalid move");
+  public getMeta(): BoardMeta {
+    const boardMeta: any[][] = Array.from({ length: this.squares.length }, () =>
+      Array.from({ length: this.size }, () => null)
+    );
+    for (let i = 0; i < this.squares.length; i++) {
+      for (let j = 0; j < this.squares[i].length; j++) {
+        const cell = this.squares[i][j];
+        const piece = cell.getPiece();
+        if (piece) {
+          const meta = piece.getMeta();
+          boardMeta[i][j] = meta;
+        }
+      }
     }
-    const fromCell: Cell = this.squares[fromX][fromY];
+    return boardMeta;
+  }
 
-    if (fromCell.isEmpty()) {
-      throw new Error("Invalid move. Selected cell is empty");
-    }
-    const fromPiece = fromCell.getPiece();
+  private mapper = {
+    [PieceType.Pawn]: Pawn,
+    [PieceType.Bishop]: Bishop,
+    [PieceType.Knight]: Knight,
+    [PieceType.Rook]: Rook,
+    [PieceType.Queen]: Queen,
+    [PieceType.King]: King,
+  };
+  private rulesMapper = {
+    [VerticalMovementRule.name]: VerticalMovementRule,
+    [HorizontalMovementRule.name]: HorizontalMovementRule,
+    [DiagonalMovementRule.name]: DiagonalMovementRule,
+  };
 
-    if (!fromCell.isEmpty() || fromPiece?.color !== color) {
-      throw new Error("Invalid move. Selected piece is not your color");
+  buildPiece(meta: PieceMeta) {
+    const c = this.mapper[meta.type as PieceType];
+    return new c(
+      meta.color,
+      meta.rules.map((ruleMeta) => {
+        const r = this.rulesMapper[ruleMeta.name];
+        return new r({ ...ruleMeta, directions: new Set(ruleMeta.directions) });
+      })
+    );
+  }
+
+  fillBoard(metas: BoardMeta) {
+    for (let row = 0; row < this.size; row++) {
+      for (let col = 0; col < this.size; col++) {
+        const meta = metas[row][col];
+
+        const cell = this.getCell(col, row);
+        if (meta) {
+          cell.putPiece(this.buildPiece(meta));
+        }
+      }
     }
-    if (!this.isMoveValid(fromPiece, fromX, fromY, toX, toY)) {
-      throw new Error("Invalid move. This piece don't have such movement rule");
-    }
-    const toCell = this.squares[toX][toY];
+  }
+  getCell(x: number, y: number) {
+    return this.squares[y][x];
+  }
+
+  updateCellsOnMove(fromCell: Cell, toCell: Cell) {
     // really move here
     if (toCell.isEmpty()) {
       const piece = fromCell.popPiece() as Piece;
@@ -56,6 +98,38 @@ export class Board {
       toCell.popPiece();
       toCell.putPiece(piece);
     }
+  }
+
+  move(color: Color, from: string, to: string) {
+    const [fromX, fromY] = fromChessToLogic(from);
+    const [toX, toY] = fromChessToLogic(to);
+
+    if (
+      !(
+        this.isIndexInValid(fromX) &&
+        this.isIndexInValid(fromY) &&
+        this.isIndexInValid(toX) &&
+        this.isIndexInValid(toY)
+      )
+    ) {
+      throw new Error("Invalid move");
+    }
+    const fromCell: Cell = this.squares[fromY][fromX];
+
+    if (fromCell.isEmpty()) {
+      throw new Error("Invalid move. Selected cell is empty");
+    }
+    const fromPiece = fromCell.getPiece();
+
+    if (fromPiece?.color !== color) {
+      throw new Error("Invalid move. Selected piece is not your color");
+    }
+    if (!this.isMoveValid(fromPiece, fromX, fromY, toX, toY)) {
+      throw new Error("Invalid move. This piece don't have such movement rule");
+    }
+    const toCell = this.squares[toY][toX];
+    // really move here
+    this.updateCellsOnMove(fromCell, toCell);
   }
 
   public isMoveValid(