+++ shared/CHANGELOG.md
@@ -1,3 +1,24 @@
+# 0.1.7
+
+Improved turn parsing
+
+# 0.1.6
+
+Fix movement rules logic
+
+# 0.1.5
+
+Fix movement rules logic
+
+# 0.1.3
+
+Fixup
+
+# 0.1.3
+
+Add metatypes for board and pieces.
+Add building from meta
+
 # 0.1.2
 
 Fix publishing scheme